+++ database/migrations/2016_10_18_100339_create_purchase_receipts_table.php
@@ -18,9 +18,9 @@ class CreatePurchaseReceiptsTable extends Migration
             $table->integer('supplier_id')->unsigned();
             $table->dateTime('datetime');
             $table->boolean('vat');
-            $table->float('total');
-            $table->float('reduction')->nullable();
-            $table->float('other_cost')->nullable();
+            $table->integer('total');
+            $table->integer('reduction')->nullable();
+            $table->integer('other_cost')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
         });
